+++ src/leaflet/overlay/graphic/index.js
@@ -3,8 +3,13 @@ import {CircleStyle, circleStyle} from './CircleStyle';
 import {Graphic, graphic} from './Graphic';
 import {ImageStyle, imageStyle} from './ImageStyle';
 
+import {GraphicCanvasRenderer} from './CanvasRenderer';
+import {GraphicWebGLRenderer} from './WebGLRenderer';
+
+export {GraphicCanvasRenderer, GraphicWebGLRenderer} ;
 
 export {CloverStyle, cloverStyle} ;
 export {CircleStyle, circleStyle} ;
 export {Graphic, graphic} ;
 export {ImageStyle, imageStyle} ;
+